+++ src/frontend/Komponenter/Behandling/Vurdering/VurderingRedigeringsmodusBaks/InnstillingTilNavKlageinstansAvsnitt.tsx
