+++ src/frontend/Komponenter/Behandling/Vurdering/VurderingRedigeringsmodusBaks/Nedtrekksliste.tsx
