+++ src/frontend/Komponenter/Behandling/Vurdering/VurderingRedigeringsmodusBaks/Tekstfelt.tsx
