+++ src/frontend/Komponenter/Behandling/Vurdering/VurderingRedigeringsmodusBaks/VurderingRedigeringsmodus.tsx
